Seperated fetching accommodations function from home
